refactor(work): extract runtime driver seam for pluggable backends
Phase 1 of the Kubernetes driver plan. Mechanics (docker CLI calls,
Engine API terminal transport, resource ownership verification) move
behind a WorkRuntimeDriver interface with DockerWorkRuntimeDriver as
the only implementation. The runtime service keeps every policy
decision: admission, leases, lifecycle locks, recovery bookkeeping,
budgets. Public exports are re-exported unchanged; zero behavior
change, full Work suite green.
